+++ ai-logic/firebase-ai/src/main/kotlin/com/google/firebase/ai/type/LiveSession.kt
@@ -435,14 +435,13 @@ internal constructor(
    * Manually marks the start of user activity, using the realtime API.
    *
    * The start of user activity is effectively the start of a user's turn, but depending on the
-   * configuration defined in [LiveRealtimeInputConfig], it may not be interpreted as an
-   * interruption. An example of the start of user activity could be the user speaking (not
-   * silence).
+   * configuration defined in [RealtimeInputConfig], it may not be interpreted as an interruption.
+   * An example of the start of user activity could be the user speaking (not silence).
    *
    * Should be followed with a call to [sendStopActivityRealtime] after all the data has been sent
    * for the user's turn.
    *
-   * Only required when automatic activity detection is disabled via [LiveRealtimeInputConfig].
+   * Only required when automatic activity detection is disabled via [RealtimeInputConfig].
    */
   public suspend fun sendStartActivityRealtime() {
     sendFrame(BidiGenerateContentRealtimeInputSetup(activityStart = true).toInternal())
@@ -456,7 +455,7 @@ internal constructor(
    *
    * Should follow after a previous call to [sendStartActivityRealtime].
    *
-   * Only required when automatic activity detection is disabled via [LiveRealtimeInputConfig].
+   * Only required when automatic activity detection is disabled via [RealtimeInputConfig].
    */
   public suspend fun sendStopActivityRealtime() {
     sendFrame(BidiGenerateContentRealtimeInputSetup(activityEnd = true).toInternal())